# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SDM Invasion-Risk API",
    "description": "Species-keyed MLOps service for MaxEnt species distribution models: incremental GBIF/iNaturalist ingestion, conditional retraining, model lineage, map-layer descriptors (titiler / pg_tileserv) and automated bulletins.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/species/search": {
      "get": {
        "tags": [
          "species"
        ],
        "summary": "Search Species",
        "description": "Resolve free text to canonical GBIF backbone taxon keys, enriched with iNat metadata.",
        "operationId": "searchSpecies",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 100,
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 25,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaxonSearchResult"
                  },
                  "title": "Response Searchspecies"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/species/{taxon_key}": {
      "get": {
        "tags": [
          "species"
        ],
        "summary": "Get Species",
        "operationId": "getSpecies",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Taxon Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpeciesDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/species/{taxon_key}/runs": {
      "post": {
        "tags": [
          "species"
        ],
        "summary": "Run Pipeline",
        "description": "Enqueue the incremental-fetch / conditional-retrain pipeline (full run if new).",
        "operationId": "runSpeciesPipeline",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Taxon Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/species/{taxon_key}/scenarios": {
      "post": {
        "tags": [
          "species"
        ],
        "summary": "Project Scenarios",
        "description": "Project the current model onto registered climate scenarios (no retraining).",
        "operationId": "projectScenarios",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Taxon Key"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/species/{taxon_key}/native-range": {
      "get": {
        "tags": [
          "species"
        ],
        "summary": "Get Native Range",
        "operationId": "getNativeRange",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Taxon Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NativeRangeOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "species"
        ],
        "summary": "Put Native Range",
        "description": "Save a user-reviewed polygon. Confirming it unlocks (re)training.",
        "operationId": "updateNativeRange",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Taxon Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NativeRangeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NativeRangeUpdateResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/species/{taxon_key}/occurrences": {
      "get": {
        "tags": [
          "species"
        ],
        "summary": "List Occurrences",
        "description": "Paginated occurrence list for table views (map layers use pg_tileserv vector tiles).",
        "operationId": "listOccurrences",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Taxon Key"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "GBIF",
                    "iNaturalist"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "range_label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "native",
                    "introduced",
                    "unknown"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Range Label"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OccurrencePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/species/{taxon_key}/occurrences/timeline": {
      "get": {
        "tags": [
          "species"
        ],
        "summary": "Occurrence Timeline",
        "operationId": "getOccurrenceTimeline",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Taxon Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Timeline"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/species/{taxon_key}/severity-config": {
      "put": {
        "tags": [
          "species"
        ],
        "summary": "Put Severity Config",
        "description": "Update the severity weights/priors and re-weight the current model's components\n(no retraining; returns null if no model exists yet).",
        "operationId": "updateSeverityConfig",
        "parameters": [
          {
            "name": "taxon_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Taxon Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeverityConfigModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SeverityResult"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Updateseverityconfig"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobKind": {
        "type": "string",
        "enum": [
          "pipeline",
          "bulletin",
          "scenarios",
          "validation"
        ],
        "title": "JobKind"
      },
      "JobOut": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "taxon_key": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taxon Key"
          },
          "kind": {
            "$ref": "#/components/schemas/JobKind"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "created_ts": {
            "type": "string",
            "format": "date-time",
            "title": "Created Ts"
          },
          "updated_ts": {
            "type": "string",
            "format": "date-time",
            "title": "Updated Ts"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "taxon_key",
          "kind",
          "status",
          "progress",
          "created_ts",
          "updated_ts"
        ],
        "title": "JobOut"
      },
...
      "NativeRangeOut": {
        "properties": {
          "taxon_key": {
            "type": "integer",
            "title": "Taxon Key"
          },
          "geometry": {
            "$ref": "#/components/schemas/GeoJSONMultiPolygon"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "confirmed"
            ],
            "title": "Status"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "updated_ts": {
            "type": "string",
            "format": "date-time",
            "title": "Updated Ts"
          },
          "confirmed_ts": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed Ts"
          }
        },
        "type": "object",
        "required": [
          "taxon_key",
          "geometry",
          "status",
          "source",
          "updated_ts"
        ],
        "title": "NativeRangeOut"
      },
...
      "RunRequest": {
        "properties": {
          "force_retrain": {
            "type": "boolean",
            "title": "Force Retrain",
            "default": false
          },
          "full_resync": {
            "type": "boolean",
            "title": "Full Resync",
            "description": "Re-fetch all records (not only the delta) and append missing ones",
            "default": false
          }
        },
        "type": "object",
        "title": "RunRequest"
      },
...
      "SpeciesDetail": {
        "properties": {
          "taxon_key": {
            "type": "integer",
            "title": "Taxon Key"
          },
          "scientific_name": {
            "type": "string",
            "title": "Scientific Name"
          },
          "common_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Common Name"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "status": {
            "$ref": "#/components/schemas/SpeciesStatus"
          },
          "n_occurrences_total": {
            "type": "integer",
            "title": "N Occurrences Total"
          },
          "n_occurrences_native": {
            "type": "integer",
            "title": "N Occurrences Native"
          },
          "n_pending_records": {
            "type": "integer",
            "title": "N Pending Records"
          },
          "model_version": {
            "type": "integer",
            "title": "Model Version"
          },
          "model_trained_ts": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Trained Ts"
          },
          "last_gbif_fetch_ts": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Gbif Fetch Ts"
          },
          "retrain_needed": {
            "type": "boolean",
            "title": "Retrain Needed"
          },
          "severity_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity Score",
            "description": "0\u2013100, current model"
          },
          "model_type": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "A",
                  "B"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Type"
          },
          "canonical_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Name"
          },
          "inat_taxon_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inat Taxon Id"
          },
          "last_gbif_download_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Gbif Download Key"
          },
          "last_gbif_download_doi": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Gbif Download Doi"
          },
          "bioclim_version_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bioclim Version Used"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "native_range_status": {
            "type": "string",
            "enum": [
              "missing",
              "draft",
              "confirmed"
            ],
            "title": "Native Range Status"
          },
          "sources": {
            "$ref": "#/components/schemas/OccurrenceSources"
          },
          "current_model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "severity_config": {
            "$ref": "#/components/schemas/SeverityConfigModel"
          }
        },
        "type": "object",
        "required": [
          "taxon_key",
          "scientific_name",
          "status",
          "n_occurrences_total",
          "n_occurrences_native",
          "n_pending_records",
          "model_version",
          "retrain_needed",
          "native_range_status",
          "sources",
          "severity_config"
        ],
        "title": "SpeciesDetail"
      },
...
      "TaxonSearchResult": {
        "properties": {
          "taxon_key": {
            "type": "integer",
            "title": "Taxon Key"
          },
          "scientific_name": {
            "type": "string",
            "title": "Scientific Name"
          },
          "canonical_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Name"
          },
          "rank": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "breadcrumb": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Breadcrumb",
            "description": "Kingdom \u2192 genus"
          },
          "common_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Common Name"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "inat_taxon_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inat Taxon Id"
          },
          "in_registry": {
            "type": "boolean",
            "title": "In Registry",
            "default": false
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpeciesStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "taxon_key",
          "scientific_name"
        ],
        "title": "TaxonSearchResult"
      },
...
    }
  }
}